fix(nx-vercel): copy the env file instead of symlinking it, to work better with caching
The Nx cache would otherwise create a directory, instead of a file, weirdly enough. But even if it
did the right thing, it would still not work with remote caching as it was an absolute symlink.

Whether env files should end up in a remote cache is a separate question, but now it's at least
possible to do.