feat(nx-vercel,nx-github,utilities): rework parameter expansion and post-targets
I realized environment variables wouldn't be a good place to store error logs from the CI, so
`deploy-prebuilt` needed another way to inform post-targets of the results. As a result executors
can now send whole JSON objects as metadata for the post-targets.

The substitution syntax has changed to support object paths and to be less ambiguous. It's now
similar to GitHub Actions expressions in appearance, but without all the fancy functions and
operators.